chore: restructure dependencies
This change moves the existing `dependencies` to `devDependencies` and
adds the following packages as `peerDependencies`:

- `@chakra-ui/react`
- `@chakra-ui/react-utils`
- `formik`
- `react`

We need the existing `dependencies` in order to build and because they're
`peerDependencies` required by `@chakra-ui/react`.

We consider the others to be `peerDependencies` because this package is
a "plugin" for Chakra, Formik, and React, in the sense that this package
only makes sense in the context of all three being installed and used
together. That means this package would always be installed _in addition
to_ them, and `peerDependencies` allows us to establish that
relationship.